// sdk/bindings/AppConnectEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DeviceMetadata } from "./DeviceMetadata";

export interface AppConnectEvent { sessionId: string, deviceMetadata: DeviceMetadata, language: string, timezone: string, network: string, newSession: boolean, }