// sdk/bindings/ChangeNetworkResolveEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RequestFail } from "./RequestFail";

export interface ChangeNetworkResolveEvent { sessionId: string, requestId: string, newNetwork: string | null, failureReason?: RequestFail, }