// sdk/bindings/ClientConnectEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionType } from "./SessionType";

export interface ClientConnectEvent { clientId: string, sessionId: string, walletName: string, walletType: string, sessionType: SessionType, }